requirements.txt
loguru==0.7.2
PyYAML==6.0
Requests==2.31.0
beautifulsoup4==4.12.3